Crate linera_witty

Linera Witty

This crate allows generating WIT files and host side code to interface with WebAssembly guests that adhere to the WIT interface format. The source of truth for the generated code and WIT files is the Rust source code.

Modules

hlist

Module that holds HList data structures, implementations, and typeclasses.

Macros

HList

Returns a type signature for an HList of the provided types

hlist

Returns an HList based on the values passed in.

Structs

FakeInstance

A fake Wasm instance.

FakeRuntime

A fake Wasm runtime.

GuestPointer

An address for a location in a guest WebAssembly module’s memory.

HNil

Represents the right-most end of a heterogeneous list

Memory

A handle to interface with a guest Wasm module instance’s memory.

Enums

RuntimeError

Errors that can occur when using a Wasm runtime.

Traits

HList

Typeclass for HList-y behaviour

InstanceWithMemory

Trait alias for a Wasm module instance with the WIT Canonical ABI functions.

Layout

Representation of the memory layout of complex types as a sequence of fundamental WIT types.

Runtime

A Wasm runtime.

RuntimeMemory

Interface for accessing a runtime specific memory.

Split

Compile time splitting of heterogeneous lists.

WitLoad

A type that can be loaded from a guest Wasm module.

WitType

A type that is representable by fundamental WIT types.

Derive Macros

WitLoad

Derives WitLoad for the Rust type.

WitType

Derives WitType for a Rust type.