Struct linera_witty::GuestPointer

pub struct GuestPointer(_);

An address for a location in a guest WebAssembly module’s memory.

Implementations

impl GuestPointer

pub const fn aligned_at(&self, alignment: u32) -> Self

Returns a new address that’s the current address advanced to add padding to ensure it’s aligned to the alignment byte boundary.

pub const fn after<T: WitType>(&self) -> Self

Returns a new address that’s the current address advanced to after the size of T.

pub const fn after_padding_for<T: WitType>(&self) -> Self

Returns a new address that’s the current address advanced to add padding to ensure it’s aligned properly for T.

pub const fn index<T: WitType>(&self, index: u32) -> Self

Returns the address of an element in a contiguous list of properly aligned T types.