Struct linera_witty::MockExportedFunction

pub struct MockExportedFunction<Parameters, Results, UserData> { /* private fields */ }

A helper type to verify how many times an exported function is called.

Implementations

impl<Parameters, Results, UserData> MockExportedFunction<Parameters, Results, UserData>where Parameters: 'static, Results: 'static, UserData: 'static,

pub fn new( name: impl Into<String>, handler: impl Fn(MockInstance<UserData>, Parameters) -> Result<Results, RuntimeError> + 'static, expected_calls: usize ) -> Self

Creates a new MockExportedFunction for the exported function with the provided name.

Every call to the exported function is called is forwarded to the handler and an internal counter is incremented. When the MockExportedFunction instance is dropped (which should be done at the end of the test}, it asserts that the function was called expected_calls times.

pub fn register(&self, instance: &mut MockInstance<UserData>)

Registers this MockExportedFunction with the mock instance.