Struct linera_witty::MockInstance

pub struct MockInstance { /* private fields */ }

A fake Wasm instance.

Only contains exports for the memory and the canonical ABI allocation functions.

Implementations

impl MockInstance

pub fn with_exported_function<Parameters, Results, Handler>( self, name: impl Into<String>, handler: Handler ) -> Selfwhere Parameters: 'static, Results: 'static, Handler: Fn(Parameters) -> Result<Results, RuntimeError> + 'static,

Adds a mock exported function to this MockInstance.

The handler will be called whenever the exported function is called.

pub fn add_exported_function<Parameters, Results, Handler>( &mut self, name: impl Into<String>, handler: Handler ) -> &mut Selfwhere Parameters: 'static, Results: 'static, Handler: Fn(Parameters) -> Result<Results, RuntimeError> + 'static,

Adds a mock exported function to this MockInstance.

The handler will be called whenever the exported function is called.