Trait linera_witty::Instance

pub trait Instance: Sized {
    type Runtime: Runtime;
    type UserData;
    type UserDataReference<'a>: Deref<Target = Self::UserData>
       where Self::UserData: 'a,
             Self: 'a;
    type UserDataMutReference<'a>: DerefMut<Target = Self::UserData>
       where Self::UserData: 'a,
             Self: 'a;

    // Required methods
    fn load_export(
        &mut self,
        name: &str,
    ) -> Option<<Self::Runtime as Runtime>::Export>;
    fn user_data(&self) -> Self::UserDataReference<'_>;
    fn user_data_mut(&mut self) -> Self::UserDataMutReference<'_>;
}

An active guest Wasm module.

Required Associated Types

type Runtime: Runtime

The runtime this instance is running in.

type UserData

Custom user data stored in the instance.

type UserDataReference<'a>: Deref<Target = Self::UserData> where Self::UserData: 'a, Self: 'a

A reference to the custom user data stored in the instance.

type UserDataMutReference<'a>: DerefMut<Target = Self::UserData> where Self::UserData: 'a, Self: 'a

A mutable reference to the custom user data stored in the instance.

Required Methods

fn load_export( &mut self, name: &str, ) -> Option<<Self::Runtime as Runtime>::Export>

Loads an export from the guest module.

fn user_data(&self) -> Self::UserDataReference<'_>

Returns a reference to the custom user data stored in this instance.

fn user_data_mut(&mut self) -> Self::UserDataMutReference<'_>

Returns a mutable reference to the custom user data stored in this instance.

Object Safety

This trait is not object safe.