linera_witty

Trait InstanceWithMemory

pub trait InstanceWithMemory: CabiReallocAlias + CabiFreeAlias {
    // Required method
    fn memory_from_export(
        &self,
        export: <Self::Runtime as Runtime>::Export,
    ) -> Result<Option<<Self::Runtime as Runtime>::Memory>, RuntimeError>;

    // Provided method
    fn memory(&mut self) -> Result<Memory<'_, Self>, RuntimeError> { ... }
}

Trait alias for a Wasm module instance with the WIT Canonical ABI functions.

Required Methods

fn memory_from_export( &self, export: <Self::Runtime as Runtime>::Export, ) -> Result<Option<<Self::Runtime as Runtime>::Memory>, RuntimeError>

Converts an export into the runtime’s specific memory type.

Provided Methods

fn memory(&mut self) -> Result<Memory<'_, Self>, RuntimeError>

Returns the memory export from the current Wasm module instance.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.