linera_witty

Trait JoinFlatLayouts

pub trait JoinFlatLayouts<Target> {
    // Required methods
    fn into_joined(self) -> Target;
    fn from_joined(joined: Target) -> Self;
}

Allows converting between the current flat layout and the joined Target flat layout, which may be longer or have some elements wider than the current elements.

Required Methods

fn into_joined(self) -> Target

Converts the current flat layout into a the joined Target flat layout.

fn from_joined(joined: Target) -> Self

Converts from the joined Target flat layout into the current flat layout.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.