Trait linera_witty::Layout

pub trait Layout: Sealed + HList {
    type Flat: FlatLayout;

    const ALIGNMENT: u32;

    // Required method
    fn flatten(self) -> Self::Flat;
}

Representation of the memory layout of complex types as a sequence of fundamental WIT types.

Required Associated Types

type Flat: FlatLayout

Result of flattening this layout.

Required Associated Constants

const ALIGNMENT: u32

The alignment boundary required for the layout.

Required Methods

fn flatten(self) -> Self::Flat

Flattens this layout into a layout consisting of native WebAssembly types.

The resulting flat layout does not have any empty items.

Object Safety

This trait is not object safe.