Trait linera_witty::Merge

pub trait Merge<Other>: Sized {
    type Output;
}

Merging of two heterogeneous lists, resulting in a new heterogenous list where every element is of type Either<Left, Right>, where Left is an element from the current list and Right is an element from the Other list.

Required Associated Types

type Output

The resulting heterogeneous list with elements of both input lists.

Object Safety

This trait is not object safe.