Trait linera_witty::Runtime

pub trait Runtime: Sized {
    type Export;
    type Memory;
}

A Wasm runtime.

Shared types between different guest instances that use the same runtime.

Required Associated Types

type Export

A handle to something exported from a guest Wasm module.

type Memory

A handle to the guest Wasm module’s memory.

Object Safety

This trait is not object safe.