Trait linera_witty::WitLoad

pub trait WitLoad: WitType {
    // Required methods
    fn load<Instance>(
        memory: &Memory<'_, Instance>,
        location: GuestPointer
    ) -> Result<Self, RuntimeError>
       where Instance: InstanceWithMemory,
             <Instance::Runtime as Runtime>::Memory: RuntimeMemory<Instance>;
    fn lift_from<Instance>(
        flat_layout: <Self::Layout as Layout>::Flat,
        memory: &Memory<'_, Instance>
    ) -> Result<Self, RuntimeError>
       where Instance: InstanceWithMemory,
             <Instance::Runtime as Runtime>::Memory: RuntimeMemory<Instance>;
}

A type that can be loaded from a guest Wasm module.

Required Methods

fn load<Instance>( memory: &Memory<'_, Instance>, location: GuestPointer ) -> Result<Self, RuntimeError>where Instance: InstanceWithMemory, <Instance::Runtime as Runtime>::Memory: RuntimeMemory<Instance>,

Loads an instance of the type from the location in the guest’s memory.

fn lift_from<Instance>( flat_layout: <Self::Layout as Layout>::Flat, memory: &Memory<'_, Instance> ) -> Result<Self, RuntimeError>where Instance: InstanceWithMemory, <Instance::Runtime as Runtime>::Memory: RuntimeMemory<Instance>,

Lifts an instance of the type from the flat_layout representation.

May read from the memory if the type has references to heap data.