linera_witty
Module wasmer

Support for the Wasmer runtime.

Structs

EntrypointInstance

Necessary data for implementing an entrypoint Instance.

Environment

A slot to store a wasmer::Instance in a way that can be shared with reentrant calls.

FunctionEnvMut

A temporary handle to a FunctionEnv.

InstanceBuilder

Helper to create Wasmer Instance implementations.

Wasmer

Representation of the Wasmer runtime.

Traits

WasmerParameters

Conversions between flat layouts and Wasmer parameter types.

WasmerResults

Conversions between flat layouts and Wasmer function result types.

Type Aliases

ReentrantInstance

Alias for the Instance implementation made available inside host functions called by the guest.