Trait linera_witty::wasmer::WasmerParameters

pub trait WasmerParameters: FlatLayout {
    type ImportParameters;
    type ExportParameters;

    // Required methods
    fn into_wasmer(self) -> Self::ImportParameters;
    fn from_wasmer(parameters: Self::ExportParameters) -> Self;
}

Conversions between flat layouts and Wasmer parameter types.

Required Associated Types

type ImportParameters

The type Wasmer uses to represent the parameters in a function imported from a guest.

type ExportParameters

The type Wasmer uses to represent the parameters in a function exported from a host.

Required Methods

fn into_wasmer(self) -> Self::ImportParameters

Converts from this flat layout into Wasmer’s representation for functions imported from a guest.

fn from_wasmer(parameters: Self::ExportParameters) -> Self

Converts from Wasmer’s representation for functions exported from the host into this flat layout.

Object Safety

This trait is not object safe.