Trait linera_witty::JoinFlatLayouts

pub trait JoinFlatLayouts<Target> {
    // Required methods
    fn into_joined(self) -> Target;
    fn from_joined(joined: Target) -> Self;
}

Allows converting between the current flat layout and the joined Target flat layout, which may be longer or have some elements wider than the current elements.

Required Methods

fn into_joined(self) -> Target

Converts the current flat layout into a the joined Target flat layout.

fn from_joined(joined: Target) -> Self

Converts from the joined Target flat layout into the current flat layout.

Object Safety

This trait is not object safe.