linera_witty

Trait Merge

pub trait Merge<Other>: Sized {
    type Output;
}

Merging of two heterogeneous lists, resulting in a new heterogeneous list where every element is of type Either<Left, Right>, where Left is an element from the current list and Right is an element from the Other list.

Required Associated Types

type Output

The resulting heterogeneous list with elements of both input lists.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.