Trait linera_witty::Split

pub trait Split<Target> {
    type Remainder;

    // Required method
    fn split(self) -> (Target, Self::Remainder);
}

Compile time splitting of heterogeneous lists.

Allows splitting a heterogeneous list at a certain point, determined by the Target type, which should match this list type until a certain element. The list after that point is returned as the Remainder type.

Required Associated Types

type Remainder

The tail of remaining elements after splitting up the list.

Required Methods

fn split(self) -> (Target, Self::Remainder)

Splits the current heterogeneous list in two.