linera_witty
Module wasmtime

Support for the Wasmtime runtime.

Re-exports

pub use anyhow;

Structs

Caller

A structure representing the caller’s context when creating a function via Func::wrap.

EntrypointInstance

Necessary data for implementing an entrypoint Instance.

Linker

Structure used to link wasm modules/instances together.

Wasmtime

Representation of the Wasmtime runtime.

Traits

WasmtimeParameters

Conversions between flat layouts and Wasmtime parameter types.

WasmtimeResults

Conversions between flat layouts and Wasmtime function result types.

Type Aliases

ReentrantInstance

Alias for the Instance implementation made available inside host functions called by the guest.